Move Version out of scala.macros
